+++ CHANGELOG.md
@@ -4,6 +4,7 @@ LUYA CHANGELOG
 1.0.0-beta7 (in progress)
 -------------------------
 
+- `#863` Fixed behavior of module controller layouts when using in cms context.
 - `#839` Styled the form search input.
 - `#856` Styled the new is Homepage indicator. Added check_circle in green.
 - `#857` Improved responsive behavior of navbar. Moved icons above text. Updated colors.